Modificacion: ahora los creditos de un usuario son de tipo double, asi como el precio de un turno, el descuento aplicado a un turno y el el balance de una tarjeta
Agregacion: ahora es posible aplicarle un descuento a los turnos cuando son confirmados

Agregacion: ahora se le devuelven los creditos a un usuario cuando su turno es cancelado o marcado como perdido

Agregacion: le agregue el permiso a los usuarios para que puedan aplicar descuentos a sus turnos

Agregacion: agrego la ruta para aplicar un descuento

+++ database/migrations/2023_06_17_215816_donation_campaigns.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('fundraising_goal');
             $table->string('description');
             $table->string('photo');
-            $table->decimal('current_fundraised', 10, 2)->unisgned()->default(0.00);
+            $table->double('current_fundraised', 10, 2)->unisgned()->default(0.00);
         });
     }
